+++ lib/async_task.dart
@@ -2,4 +2,5 @@
 library async_task;
 
 export 'src/async_task_base.dart';
+export 'src/async_task_channel.dart';
 export 'src/async_task_shared_data.dart';